reshuffle before tests
