+++ .gitignore
@@ -1,0 +1,6 @@
+__pycache__
+*.pyc
+.idea
+.vscode
+.pytest_cache
+*.egg-info/
