--- pyroads/__init__.py
@@ -1,1 +1,0 @@
-#Initialise mainroads package
