+++ setup.cfg
@@ -1,0 +1,14 @@
+[metadata]
+name = pyroads
+version = 0.2.0
+author = Shaan Ciantar
+description = For analysis of road asset data
+url = https://github.com/shaan-nmb/pyroads
+
+[options]
+package_dir =
+    = src
+packages =
+    pyroads
+install_requires =
+    pandas
