+++ src/pyroads/mappings/as_meters.py
@@ -1,0 +1,5 @@
+def as_metres(var):
+
+    m = round(var*1000).astype(int)
+
+    return m
