+++ test/test_stretch.py
@@ -1,0 +1,5 @@
+
+from pyroads.reshape.stretch import stretch
+
+def test_tests():
+	assert True
